Fix errors in previous commit

+++ modules/core/m_message.c
@@ -39,6 +39,7 @@
 #include "irc_string.h"
 #include "hash.h"
 #include "packet.h"
+#include "list.h"
 
 struct entity
 {
@@ -637,7 +638,7 @@ msg_client(int p_or_n, const char *command, struct Client *source_p,
     if (source_p != target_p && !accept_message(target_p, source_p) &&
         !(IsOper(target_p) && ConfigFileEntry.opers_bypass_callerid == 1))
     {
-      if (dlink_list_length(&source_p->allow_list) < ConfigFileEntry.max_accept)
+      if (dlink_list_length(&source_p->allow_list) < (unsigned long)ConfigFileEntry.max_accept)
       {
         dlinkAdd(target_p, make_dlink_node(), &source_p->allow_list);
         dlinkAdd(source_p, make_dlink_node(), &target_p->on_allow_list);
